Safari-611.1.10/LayoutTests/js/dom/Promise-init-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/Promise-init-in-workers.js
PASS [Worker] promise instanceof Promise is true
PASS [Worker] promise.constructor is Promise
PASS [Worker] thisInInit === promise is false
PASS [Worker] thisInInit === global is true
PASS [Worker] resolve instanceof Function is true
PASS [Worker] reject instanceof Function is true
PASS [Worker] new Promise() threw exception TypeError: Promise constructor takes a function argument.
PASS [Worker] new Promise(37) threw exception TypeError: Promise constructor takes a function argument.
PASS [Worker] promise = new Promise(function() { throw Error("foo"); }) did not throw exception.
PASS [Worker] result.message is "foo"
PASS [Worker] fulfilled
PASS [Worker] result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
